+++ Sources/MapboxCoreNavigation/PredictiveCache/NativeNavigatorProvider.swift
@@ -1,0 +1,11 @@
+import MapboxNavigationNative
+
+protocol NativeNavigatorProvider {
+    static var navigator: MapboxNavigationNative.Navigator { get }
+}
+
+extension Navigator: NativeNavigatorProvider {
+    static var navigator: MapboxNavigationNative.Navigator {
+        return Self.shared.navigator
+    }
+}